+++ .github/workflows/ci-rust.yml
@@ -14,7 +14,6 @@ name: Rust CI
 env:
   RUNTIME_CRATE: extism-runtime
   LIBEXTISM_CRATE: libextism
-  RUST_SDK_CRATE: extism
 
 jobs:
   lib:
@@ -86,7 +85,11 @@ jobs:
       - name: Lint
         run: cargo clippy --release --all-features --no-deps -p ${{ env.RUNTIME_CRATE }}
       - name: Test
+        run: cargo test --release -p ${{ env.RUNTIME_CRATE }}
+      - name: Test all features
         run: cargo test --all-features --release -p ${{ env.RUNTIME_CRATE }}
+      - name: Test no features
+        run: cargo test --no-default-features --release -p ${{ env.RUNTIME_CRATE }}
 
   rust:
     name: Rust
