+++ .github/workflows/release-rust.yaml
@@ -26,14 +26,12 @@ jobs:
           # order of crate publication matter: manifest, runtime, rust
           cargo publish --manifest-path manifest/Cargo.toml
           # allow for crates.io to update so dependant crates can locate extism-manifest
-          sleep 5 
+          sleep 10 
 
-      - name: Release Rust Host SDK
+      - name: Release Runtime
         env:
           CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_API_TOKEN }}
         run: |
-
           cargo publish --manifest-path runtime/Cargo.toml --no-verify
-          cargo publish --manifest-path rust/Cargo.toml
