+++ manifest/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "extism-manifest"
-version = "0.5.0"
+version = "1.0.0-alpha.0"
 edition = "2021"
 authors = ["The Extism Authors", "oss@extism.org"]
 license = "BSD-3-Clause"
