+++ runtime/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "extism-runtime"
-version = "0.5.0"
+name = "extism"
+version = "1.0.0-alpha.0"
 edition = "2021"
 authors = ["The Extism Authors", "oss@extism.org"]
 license = "BSD-3-Clause"
 homepage = "https://extism.org"
 repository = "https://github.com/extism/extism"
-description = "Extism runtime component"
+description = "Extism runtime and Rust SDK"
 
 [dependencies]
 wasmtime = ">= 10.0.0, < 12.0.0"
@@ -22,7 +22,7 @@ log4rs = "1.1"
 url = "2"
 glob = "0.3"
 ureq = {version = "2.5", optional=true}
-extism-manifest = { version = "0.5.0", path = "../manifest" }
+extism-manifest = { version = "1.0.0-alpha.0", path = "../manifest" }
 uuid = { version = "1", features = ["v4"] }
 libc = "0.2"
