+++ runtime/build.rs
@@ -14,10 +14,9 @@ fn main() {
         .with_pragma_once(true)
         .with_after_include(fn_macro)
         .rename_item("Size", "ExtismSize")
-        .rename_item("PluginIndex", "ExtismPlugin")
         .rename_item("ValType", "ExtismValType")
         .rename_item("ValUnion", "ExtismValUnion")
-        .rename_item("Internal", "ExtismCurrentPlugin")
+        .rename_item("CurrentPlugin", "ExtismCurrentPlugin")
         .rename_item("Plugin", "ExtismPlugin")
         .rename_item("Function", "ExtismFunction")
         .with_style(cbindgen::Style::Type)
