+++ runtime/src/pdk.rs
@@ -22,18 +22,18 @@ macro_rules! args {
 /// Params: i64 (offset)
 /// Returns: i64 (offset)
 pub(crate) fn config_get(
-    mut caller: Caller<Internal>,
+    mut caller: Caller<CurrentPlugin>,
     input: &[Val],
     output: &mut [Val],
 ) -> Result<(), Error> {
-    let data: &mut Internal = caller.data_mut();
+    let data: &mut CurrentPlugin = caller.data_mut();
 
     let offset = args!(input, 0, i64) as u64;
     let key = data.memory_read_str(offset)?;
     let key = unsafe {
         std::str::from_utf8_unchecked(std::slice::from_raw_parts(key.as_ptr(), key.len()))
     };
-    let val = data.internal().manifest.config.get(key);
+    let val = data.manifest.config.get(key);
     let ptr = val.map(|x| (x.len(), x.as_ptr()));
     let mem = match ptr {
         Some((len, ptr)) => {
@@ -53,18 +53,18 @@ pub(crate) fn config_get(
 /// Params: i64 (offset)
 /// Returns: i64 (offset)
 pub(crate) fn var_get(
-    mut caller: Caller<Internal>,
+    mut caller: Caller<CurrentPlugin>,
     input: &[Val],
     output: &mut [Val],
 ) -> Result<(), Error> {
-    let data: &mut Internal = caller.data_mut();
+    let data: &mut CurrentPlugin = caller.data_mut();
 
     let offset = args!(input, 0, i64) as u64;
     let key = data.memory_read_str(offset)?;
     let key = unsafe {
         std::str::from_utf8_unchecked(std::slice::from_raw_parts(key.as_ptr(), key.len()))
     };
-    let val = data.internal().vars.get(key);
+    let val = data.vars.get(key);
     let ptr = val.map(|x| (x.len(), x.as_ptr()));
     let mem = match ptr {
         Some((len, ptr)) => {
@@ -84,11 +84,11 @@ pub(crate) fn var_get(
 /// Params: i64 (key offset), i64 (value offset)
 /// Returns: none
 pub(crate) fn var_set(
-    mut caller: Caller<Internal>,
+    mut caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
-    let data: &mut Internal = caller.data_mut();
+    let data: &mut CurrentPlugin = caller.data_mut();
 
     let mut size = 0;
     for v in data.vars.values() {
@@ -129,7 +129,7 @@ pub(crate) fn var_set(
 /// Params: i64 (offset to JSON encoded HttpRequest), i64 (offset to body or 0)
 /// Returns: i64 (offset)
 pub(crate) fn http_request(
-    #[allow(unused_mut)] mut caller: Caller<Internal>,
+    #[allow(unused_mut)] mut caller: Caller<CurrentPlugin>,
     input: &[Val],
     output: &mut [Val],
 ) -> Result<(), Error> {
@@ -145,7 +145,7 @@ pub(crate) fn http_request(
     #[cfg(feature = "http")]
     {
         use std::io::Read;
-        let data: &mut Internal = caller.data_mut();
+        let data: &mut CurrentPlugin = caller.data_mut();
         let http_req_offset = args!(input, 0, i64) as u64;
 
         let http_req_len = data.memory_length(http_req_offset);
@@ -158,7 +158,7 @@ pub(crate) fn http_request(
             Ok(u) => u,
             Err(e) => return Err(Error::msg(format!("Invalid URL: {e:?}"))),
         };
-        let allowed_hosts = &data.internal().manifest.allowed_hosts;
+        let allowed_hosts = &data.manifest.allowed_hosts;
         let host_str = url.host_str().unwrap_or_default();
         let host_matches = if let Some(allowed_hosts) = allowed_hosts {
             allowed_hosts.iter().any(|url| {
@@ -229,22 +229,22 @@ pub(crate) fn http_request(
 /// Params: none
 /// Returns: i32 (status code)
 pub(crate) fn http_status_code(
-    mut caller: Caller<Internal>,
+    mut caller: Caller<CurrentPlugin>,
     _input: &[Val],
     output: &mut [Val],
 ) -> Result<(), Error> {
-    let data: &mut Internal = caller.data_mut();
+    let data: &mut CurrentPlugin = caller.data_mut();
     output[0] = Val::I32(data.http_status as i32);
     Ok(())
 }
 
 pub fn log(
     level: log::Level,
-    mut caller: Caller<Internal>,
+    mut caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
-    let data: &mut Internal = caller.data_mut();
+    let data: &mut CurrentPlugin = caller.data_mut();
     let offset = args!(input, 0, i64) as u64;
     let buf = data.memory_read_str(offset);
 
@@ -259,7 +259,7 @@ pub fn log(
 /// Params: i64 (offset)
 /// Returns: none
 pub(crate) fn log_warn(
-    caller: Caller<Internal>,
+    caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
@@ -270,7 +270,7 @@ pub(crate) fn log_warn(
 /// Params: i64 (offset)
 /// Returns: none
 pub(crate) fn log_info(
-    caller: Caller<Internal>,
+    caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
@@ -281,7 +281,7 @@ pub(crate) fn log_info(
 /// Params: i64 (offset)
 /// Returns: none
 pub(crate) fn log_debug(
-    caller: Caller<Internal>,
+    caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
@@ -292,7 +292,7 @@ pub(crate) fn log_debug(
 /// Params: i64 (offset)
 /// Returns: none
 pub(crate) fn log_error(
-    caller: Caller<Internal>,
+    caller: Caller<CurrentPlugin>,
     input: &[Val],
     _output: &mut [Val],
 ) -> Result<(), Error> {
