--- rust/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "extism"
-version = "0.5.0"
-edition = "2021"
-authors = ["The Extism Authors", "oss@extism.org"]
-license = "BSD-3-Clause"
-homepage = "https://extism.org"
-repository = "https://github.com/extism/extism"
-description = "Extism Host SDK for Rust"
-
-[dependencies]
-extism-manifest = { version = "0.5.0", path = "../manifest" }
-extism-runtime = { version = "0.5.0", path = "../runtime"}
-serde_json = "1"
-log = "0.4"
-anyhow = "1"
-uuid = { version = "1", features = ["v4"] }
