--- rust/Makefile
@@ -1,2 +1,0 @@
-bindings:
-	bindgen ../runtime/extism.h --allowlist-function extism.* > src/bindings.rs
